Changed camera to free rotation and added zoom
For keyboard and mouse, move the mouse to rotate camera up down left or right and use the scrollwheel for zoom.

For joystick, use right joystick to rotate camera up down left and right and press down R3 and move up and down to zoom in and out (THIS HAS NOT BEEN TESTED).

Constants were changed for the new script so that the zoom goes out further.